Reimported Claw as fbx format.
Claw originally imported from Blender in .obj format. While in Unity from this format, the model had an extra "default" object layer that made data harder to access from scripts. So went back to Blender. (Also cleaned up the claw image a bit.) And exported it as .fbx, which is the format that Unity seems to prefer.