Added more TRACE_COLLISIONS
Started in to put in code to detect all possible interactions between objects. It is getting to be way too much and would require special test code to see if some of these events even ever happen. Just capture mid-task snapshot and will next go back to trim a lot of this out before just concentrating on the interactions i really care about.